Moved rollbar.go into its own package.
